more refactoring task chaining fixes
