+++ src/fullctl/django/models/__init__.py
@@ -1,2 +1,3 @@
 from fullctl.django.models.concrete.account import *  # noqa: F401, F403
 from fullctl.django.models.concrete.auditlog import *  # noqa: F401, F403
+from fullctl.django.models.concrete.tasks import *  # noqa: F401, F403
